+++ execution/market_test.go
@@ -499,10 +499,7 @@ func getMarketWithDP(closingAt time.Time, pMonitorSettings *types.PriceMonitorin
 				},
 			},
 		},
-		OpeningAuction: openingAuctionDuration,
-		TradingModeConfig: &types.MarketContinuous{
-			Continuous: &types.ContinuousTrading{},
-		},
+		OpeningAuction:          openingAuctionDuration,
 		PriceMonitoringSettings: pMonitorSettings,
 		LiquidityMonitoringParameters: &types.LiquidityMonitoringParameters{
 			TargetStakeParameters: &types.TargetStakeParameters{
@@ -1009,9 +1006,6 @@ func TestSetMarketID(t *testing.T) {
 						},
 					},
 				},
-			},
-			TradingModeConfig: &types.MarketContinuous{
-				Continuous: &types.ContinuousTrading{},
 			},
 		}